Added more endpoints, status indicators, and removed a LOT of dead code
